# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "main方法",
        "contact": {}
    },
    "paths": {
        "/dex/liquidity/findAllTokens": {
            "get": {
                "description": "拿所有的代币信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "拿所有的代币信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dex/liquidity/getAnotherTokenInfoByTokenId": {
            "get": {
                "description": "通过代币id拿 与之相对的 另外一个代币的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "通过代币id拿 与之相对的 另外一个代币的信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tokenId",
                        "name": "tokenId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dex/liquidity/getMatchAmount": {
            "get": {
                "description": "拿代币对信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "拿代币对信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tokenId0",
                        "name": "tokenId0",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tokenId1",
                        "name": "tokenId1",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/helloworld": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}